backend/requirements.txt
flask
google-generativeai>=0.4.0
PyMuPDF
flask-cors
